+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 ## in development:
 * **New Features:**
-    * export / import sprite-local custom block definitions, under construction
+    * export script (including dependencies)
+    * export / import sprite-local custom block definitions
     * added "combinations" primitive to the palette
 * **Notable Changes:**
     * moved "append", "reshape", "combinations" blocks down one group in the palette
@@ -15,6 +16,7 @@
 
 ### 2022-03-15
 * blocks, store, gui: deserialize new format for exported scripts
+* blocks: new "export script" feature (including dependencies)
 
 ### 2022-03-14
 * gui, byob: refactored library serialization